+++ src/main/java/edu/ie3/datamodel/io/naming/DataSourceMetaInformation.java
@@ -3,16 +3,16 @@
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
-package edu.ie3.datamodel.io.csv;
+package edu.ie3.datamodel.io.naming;
 
 import java.util.Objects;
 import java.util.UUID;
 
-/** Meta information, that can be derived from a certain file name */
-public abstract class FileNameMetaInformation {
+/** Meta information, that can be derived from a data source */
+public abstract class DataSourceMetaInformation {
   private final UUID uuid;
 
-  protected FileNameMetaInformation(UUID uuid) {
+  protected DataSourceMetaInformation(UUID uuid) {
     this.uuid = uuid;
   }
 
@@ -23,7 +23,7 @@ public abstract class FileNameMetaInformation {
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
-    if (!(o instanceof FileNameMetaInformation that)) return false;
+    if (!(o instanceof DataSourceMetaInformation that)) return false;
     return uuid.equals(that.uuid);
   }
 
@@ -34,6 +34,6 @@ public abstract class FileNameMetaInformation {
 
   @Override
   public String toString() {
-    return "FileNameMetaInformation{" + "uuid=" + uuid + '}';
+    return "DataSourceMetaInformation{" + "uuid=" + uuid + '}';
   }
 }